Добавляет const.js
