+++ src/main.js
@@ -1,5 +1,4 @@
 import FilterView from './view/list-filter-view.js';
-import SortView from './view/list-sort-view.js';
 import {render} from './render.js';
 import BoardPresenter from './presenter/board-presenter.js';
 import PointsModel from './model/points-model.js';
@@ -11,6 +10,5 @@ const pointsModel = new PointsModel();
 const boardPresenter = new BoardPresenter({boardContainer: tripEventsElement, pointsModel});
 
 render(new FilterView(), tripFiltersElement);
-render(new SortView(), tripEventsElement);
 
 boardPresenter.init();
