+++ src/presenter/board-presenter.js
@@ -1,6 +1,7 @@
 import EventsItemView from '../view/events-item-view.js';
 import EventsListView from '../view/events-list-view.js';
 import NewPointView from '../view/new-point-view.js';
+import SortView from '../view/list-sort-view.js';
 import {render} from '../render.js';
 
 export default class BoardPresenter {
@@ -15,6 +16,7 @@ export default class BoardPresenter {
     this.boardPoints = [...this.pointsModel.getPoints()];
     console.log(this.boardPoints);
 
+    render(new SortView, this.boardContainer);
     render(this.boardComponent, this.boardContainer);
     render(new NewPointView, this.boardComponent.getElement());
